Source: Do not limit hash table to 1GB size
In the soon future 1GB hash will simply not be enough with a lot
of threads. Remove this restriction

+++ Source/pvtable.h
@@ -15,6 +15,6 @@ int read_hash_entry(board_t *board, tt_t *hash_table, int alpha, int *move,
                     int beta, int depth);
 void write_hash_entry(board_t *board, tt_t *hash_table, int score, int depth,
                       int move, int hash_flag);
-void init_hash_table(engine_t *engine, tt_t *hash_table, int mb);
+void init_hash_table(engine_t *engine, tt_t *hash_table, uint64_t mb);
 uint64_t generate_hash_key(engine_t *engine, board_t *board);
 #endif
